Fix resolve-missing for records and types
This also removes an assumption that capitalized names belong to
classes.  This prevented the resolution of e.g. schemas from
prismatic/schema, which are vars that conventionally start with capital
letters.

The return value of the op has changed.  We're now dropping :type because
it no longer contains the resolution we need on the client side.  We have
to be able determine the difference between a missing:

1) record or type, as created by defrecord or deftype, respectively
2) class
3) ns

The new return value is thus an alist of ((candidate1 . type1) (candidate
2 . type2) ...), where type is in #{:type :ns :class}.

A :type is generated by both defrecord or deftype becaues the client will
do the same in both cases.

Interfaces are included under :class because info-clj doesn't make the
distinction and the client doesn't have to branch on the information.

It's quite convenient to simply read in result alist on the elisp
side, but I imagine this might not be the case in clients written in other
languages.  If we ever get other consumers it should be easy enough to
change the representation, but I don't want to write that code now so I'm
hoping for YAGNI :)