[Fix #24]: Error when we can't build AST
Previously we would just carry on when encountering failures.  At first I
added an `:ignore-errors` option to `ns-ast`, but I quickly became
convinced that this isn't the right thing to do.

A quick grep over the project revealed that `ns-ast` was in use for
clean-ns and the various functions used to find stuff.  I think we should
fail in all these cases instead of giving the illusion of a complete
result set and let the user re-run the op once the project is in a state
good enough to build an AST.