[Fix #42] Let errors bubble up to the user
Cider has a nice way of displaying stacktraces which we can use to get
better bug reports in the event of unexpected failure.

Expected exceptions aren't handled in this manner, like when we can't
build an ast because the file is in a bad state, but like before so we
can display a helpful error message instead of a stacktrace.