Refactor nrepl code into middleware.clj
The nrepl stuff belongs at the outer edges of the library.  This code
was presently riddled all over the lib.

I've removed the content form the descriptors because it was a pain to
keep these in sync with the readme.  I'd rather have one source of this
information.  I also can't imagine anyone consuming operator
documentation using the descriptor rather than the code itself or the
README.

I've removed all the middleware handlers except one: there's now a
single handler `wrap-refactor` which is capable of handling multiple ops.