[#Fix 57] <! gets removed by clean-ns
<! is a function and is removed by the surrounding go macro so it
doesn't show up in the AST. As it's not a macro we're not checking for
it in the 'macro pass' either.

In addition to checking if stuff appears in the ast we now walk each
form in the file and collect symbols if that form is going to get macro
expanded.  This solution is weak to false positives, but I'd rather
leave too much stuff in the ns than too little.