clean-ns shouldn't rewrite to use meta style
There are two ways to add a docstring to a namespace:

(ns my-ns ^{:doc "My docstring"})

or (ns my-ns "My docstring").

The choice among these is purely stylistic and we shouldn't rewrite to
favor one over the other.

In addition to preventing this rewrite we now print each kv pair of the
map on its own line.