Catch failures when updating resolve-missing cache
After adding core.async as a dependency and attempting to hotload it I
got a confusing ClassNotfoundException.  There's nothing we can do
about this situation, but we shouldn't freak the user out either when
the consequences are so trivial: resolve-missing, and
cljr-add-missing-libspec, won't work until the user has manually
loaded one of the namespaces in the new dependency.