Rewrite clean-ns
The previous implementation relied on the AST to to find some symbols
and a manual walk to find the rest.  Due to the effect of macros we're
already walking so much of the file that we might as well abandon the
AST and walk everything. The new implementation thus walks every form
in the body and collects symbols to use when pruning libspecs and
imports.

The benefit of this approach are as follows

- We're for sure not going to remove a libspec or a referred symbol
  which is needed.
- Since we don't rely on an AST the op will work on files that are in a
  'bad state'.
- This approach should be a lot easier to adapt to .cljs and .cljc files.

The disadvantage of this approach is that we're going to leave some
symbols and libspecs around in the ns form when they are unused.  This
should be limited to symbols found in the :refer clause, and the
community seems to be moving away from referring symbols in favor of
aliasing.

A future commit will take locals into account to limit the amount of
false positives we get when we're looking for a symbol.  This will
improve the accuracy if this op, but I don't think it'll ever get
perfect.

This fixes clojure-emacs/clj-refactor.el#186