[Fix #103] Prefix forms using lists
These are allowed even though they are unidiomatic.  Still, we should
support them, and rewrite the code so the user doesn't have to
manually spend time adhering to convention.

This used to work, but I introduced a regression when I rewrote the
ns-parser and eliminated the instaparse dependency.

The tests have been updated to guard against this regression happening
again.