Make clean-ns aware of flags in libspecs
Previously these flags were just dropped on the floor, but they're now
handled properly.  There's also code in place to eliminate redundant
flags, i.e. if you have :reload-all we'll remove :reload (which is
implied).

What remains now, in terms of flags, is support for (ns ... (:require ...
:reload-all)) type flags where the flags are applied to all libspecs in
the :require or :use form.  I doubt anyone is going to miss this, so I'll
just leave this out for now.