Improve error reporting in find_macros
The previous error message, from the lispreader, only included what the
problem was but gave no indication as to where we might look to fix it.

Changing to tools.reader improves the situation slightly, as it includes
line and column number, but we have to manually add the file information
to the exception to get everything we need to actually fix the issue.