Don't prune cljsjs namespaces in clean-ns
These namespaces represent externs.  For now the exact symbols are
unknowable, so we can't determine if the ns is in use.  As a temporary
fix we just ignore these libspecs.  While I consider this a temp hack
it's unlikely to ever get improved on, unless someone else builds tooling
we can delegate to.

This fixes clojure-emacs/clj-refactor.el#275