[Fix #139] clean-ns vs shorthand metadata
Metadata of type ^:whatever was pruned by clean-ns.  I had no idea this
form was legal in the ns form.  The documentation only talks about
metadata maps so this might be an implementation detail.  As long as it
doesn't cause any problems for us we should support it, as people seem
to be using this in on real projects.