Replace (reduce concat ,,,) with (mapcat ,,,)
The (reduce concat ,,,) pattern was causing the stack to overflow, this
was a bit tricky to debug, the symptom was that CIDER would just block
forever.

Also factored out a use of `proxy` so the class generation only needs to
happen once. All in all this version seems to perform significantly
better than the previous one.