Handle NoClassDefFoundError when determining class type
When `resolve-missing` gets a number of candidate classes, it will load them to
determine the type (Clojure type/record vs regular Java class). However a class
on the classpath may be compiled with dependencies that are not currently
present, causing the loading to fail.

This would cause the whole operation to fail, instead this change handles the
exception, takes it as a sign that it's a regular Java class, and continues.

For example, I do a `resolve-missing` on the symbol URL, this yields two
candidates: `java.net.URL`, and
`com.gargoylesoftware.htmlunit.javascript.host.URL`. The latter depends on
`org.w3c.dom.ElementTraversal`, which is not currently present, causing a
`NoClassDefFoundError`.