Add config for ignoring specific paths
This change will allow users to configure refactor-nrepl to ignore
certain paths when finding the dirs on classpath by specifying a seq of
regexes which the paths will be matched against.

Currently refactor-nrepl considers all the .clj files that are reachable
from project root. This is a problem if the project repository includes
example .clj files, if a user has symlinked other clj projects under
'checkouts' or if there are any other files that need to be ignored.

This change is backward compatible and wouldn't affect the current
behaviour in existing installations as the default value is an empty
seq.

Also added tests and updated README and Changelog.