[Fix #97] Use `fake.class.path` in case the project is managed by boot
Intend to work around problems discussed in boot-clj/boot#251 and
boot-clj/boot#249. Because of `boot`-s inner workings the original
directories for the project source paths are on `fake.class.path` and
the real classpath points to a temporary directory for the source
files. This confuses cljr. See corresponding fix for cider: clojure-emacs/cider-nrepl#422