[Fix #232] Add compatibility with nREPL 0.4
Support with tools.nrepl is kept for now with the assumption
that tools.nrepl namespaces would be loaded automatically by
`lein repl` or `boot repl`.

Support for tools.nrepl will likely be removed after a couple of
releases (once mainstream Lein and Boot switch to nREPL 0.4).

This also updates the archaic piggieback dep the project was using.