Upgrade mranderson to 0.5.0
mranderson 0.5.x specific config added to the project file as
well. Using resolved tree mode. This means that
mranderson works with a resolved dependency tree and does not create a
deeply nested directory structure, only adds prefixes to all
dependencies. Handles all depedendencies as first level ones and
processes them in a topological order.

Add project prefix in the project file and other config for unresolved
tree mode in case we want to switch to that in the near future