[clj-refactor.el #443] Add support for js literal
As described in
https://github.com/clojure-emacs/clj-refactor.el/issues/443, it was
impossible to use clj-clean-ns when a js literal was present:

'No reader function for tag js'

One possible solution was to use *default-data-reader-fn* with the
function `(fn [_tag value] value)`. However this had the potential of
swallowing useful error messages when dealing with other undefined
tags.

The chosen solution is to provide a dummy reader function specifically
for the #js tag (and only for the :cljs dialect). Rather than
transforming the provided form into a JS object, it returns the
clojure form, which can then be analyzed normally.