Fix artifact list cache not working
The content of the artifact cache was stored in a sorted map and then
written to disk.  The default readeres for edn does not contain a reader
capable of handling sorted maps.  This caused the artifact cache to stop
working.

This commits drops the use of sorted maps, as we never rely on them
being sorted anyway.