Bring hotload-dependency back
Previously this op was built on top of alembic, but that lib no longer
seems to be maintained at all so we're moving to pomegranate.

`tools.deps` would've been another alternative by their `add-lib`
operation has yet to be merged to the main branch.  Once it's merged and
has proven itself in the wild nothing prevents us from moving over,
though, if that solution proves to be better in some way.