[Fix #305] No newline after :as or :refer
This makes the ns form take up less vertical space and makes it easier
on the eyes.

This was never done intentionally but a side-effect of simply delegating
the printing of regular libspecs to `clojure.core.pprint` which puts
each element of a vector on its own line when the line gets long enough.