Don't suggest alias fragments with leading number
Namespace aliases are symbols, and symbols begin with a non-numeric character,
so we shouldn't suggest aliases that start with numbers.
This fixes clj-refactor.el's cljr-slash for projects that contain ns'es like
`db.migration.25`. Previously, it failed in parseclj when that tried to parse
the nrepl response containing suggestions like `{25 [db.migration.25]}`, since
`25` isn't a valid symbol.