org.clojure/tools.namespace update
It was eight releases behind compared to the latest release.

With the upgrade of org.clojure/tools.namespace some of the tests,
failed. This was because ClojureScript files weren't
touched/considered until org.clojure/tools.namespace 1.2.0 version,
when this bug was fixed and ClojureScript files started to be
considered.

I've also revised some of the tests.

Previously the tests just counted the names of the files touched,
whereas with this test it checks explicitly for the old & new file
names, as well as the names of all the files that referenced &
reference the new namespace.

And to be more explicit, originally when this feature was added, the
idea was to be able to test if a file move will update the Clojure
files which reference the file's namespace as well as any
ClojureScript files that reference it too. I believe that this test
shows that.